@@ -13,7 +13,7 @@ import {
 
 const SOURCE = 'tcomb-form-native';
 const nooptions = Object.freeze({});
-const noop = function () {};
+const noop = function() {};
 const noobj = Object.freeze({});
 const noarr = Object.freeze([]);
 const Nil = t.Nil;
@@ -27,22 +27,20 @@ function getFormComponent(type, options) {
   }
   const name = t.getTypeName(type);
   switch (type.meta.kind) {
-    case 'irreducible' :
-      return (
-        type === t.Boolean ? Checkbox :
-        type === t.Date ?  DatePicker :
-                          Textbox
-      );
-    case 'struct' :
+    case 'irreducible':
+      return type === t.Boolean
+        ? Checkbox
+        : type === t.Date ? DatePicker : Textbox;
+    case 'struct':
       return Struct;
-    case 'list' :
+    case 'list':
       return List;
-    case 'enums' :
+    case 'enums':
       return Select;
-    case 'maybe' :
-    case 'subtype' :
+    case 'maybe':
+    case 'subtype':
       return getFormComponent(type.meta.type, options);
-    default :
+    default:
       t.fail(`[${SOURCE}] unsupported type ${name}`);
   }
 }
@@ -59,7 +57,6 @@ function getComparator(order) {
 }
 
 class Component extends React.Component {
-
   constructor(props) {
     super(props);
     this.typeInfo = getTypeInfo(props.type);
@@ -74,12 +71,11 @@ class Component extends React.Component {
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    const should = (
+    const should =
       nextState.value !== this.state.value ||
       nextState.hasError !== this.state.hasError ||
       nextProps.options !== this.props.options ||
-      nextProps.type !== this.props.type
-    );
+      nextProps.type !== this.props.type;
     return should;
   }
 
@@ -87,17 +83,22 @@ class Component extends React.Component {
     if (props.type !== this.props.type) {
       this.typeInfo = getTypeInfo(props.type);
     }
-    this.setState({value: this.getTransformer().format(props.value)});
+    this.setState({ value: this.getTransformer().format(props.value) });
   }
 
   onChange(value) {
-    this.setState({value}, () => this.props.onChange(value, this.props.ctx.path));
+    this.setState({ value }, () =>
+      this.props.onChange(value, this.props.ctx.path)
+    );
   }
 
   getValidationOptions() {
     return {
       path: this.props.ctx.path,
-      context: t.mixin(t.mixin({}, this.props.context || this.props.ctx.context), { options: this.props.options })
+      context: t.mixin(
+        t.mixin({}, this.props.context || this.props.ctx.context),
+        { options: this.props.options }
+      )
     };
   }
 
@@ -110,16 +111,20 @@ class Component extends React.Component {
   }
 
   removeErrors() {
-    this.setState({hasError: false});
+    this.setState({ hasError: false });
   }
 
   pureValidate() {
-    return t.validate(this.getValue(), this.props.type, this.getValidationOptions());
+    return t.validate(
+      this.getValue(),
+      this.props.type,
+      this.getValidationOptions()
+    );
   }
 
   validate() {
     const result = this.pureValidate();
-    this.setState({hasError: !result.isValid()});
+    this.setState({ hasError: !result.isValid() });
     return result;
   }
 
@@ -134,7 +139,12 @@ class Component extends React.Component {
   getDefaultLabel() {
     const ctx = this.props.ctx;
     if (ctx.label) {
-      return ctx.label + (this.typeInfo.isMaybe ? this.getI18n().optional : this.getI18n().required);
+      return (
+        ctx.label +
+        (this.typeInfo.isMaybe
+          ? this.getI18n().optional
+          : this.getI18n().required)
+      );
     }
   }
 
@@ -148,10 +158,15 @@ class Component extends React.Component {
 
   getError() {
     if (this.hasError()) {
-      const error = this.props.options.error || this.typeInfo.getValidationErrorMessage;
+      const error =
+        this.props.options.error || this.typeInfo.getValidationErrorMessage;
       if (t.Function.is(error)) {
         const validationOptions = this.getValidationOptions();
-        return error(this.getValue(), validationOptions.path, validationOptions.context);
+        return error(
+          this.getValue(),
+          validationOptions.path,
+          validationOptions.context
+        );
       }
       return error;
     }
@@ -186,35 +201,41 @@ class Component extends React.Component {
   render() {
     const locals = this.getLocals();
     // getTemplate is the only required implementation when extending Component
-    t.assert(t.Function.is(this.getTemplate), `[${SOURCE}] missing getTemplate method of component ${this.constructor.name}`);
+    t.assert(
+      t.Function.is(this.getTemplate),
+      `[${SOURCE}] missing getTemplate method of component ${this.constructor
+        .name}`
+    );
     const template = this.getTemplate();
     return template(locals);
   }
-
 }
 
 Component.transformer = {
-  format: value => Nil.is(value) ? null : value,
+  format: value => (Nil.is(value) ? null : value),
   parse: value => value
 };
 
 function toNull(value) {
-  return (t.String.is(value) && value.trim() === '') || Nil.is(value) ? null : value;
+  return (t.String.is(value) && value.trim() === '') || Nil.is(value)
+    ? null
+    : value;
 }
 
 function parseNumber(value) {
   const n = parseFloat(value);
-  const isNumeric = (value - n + 1) >= 0;
+  const isNumeric = value - n + 1 >= 0;
   return isNumeric ? n : toNull(value);
 }
 
 class Textbox extends Component {
-
   getTransformer() {
     const options = this.props.options;
-    return options.transformer ? options.transformer :
-      this.typeInfo.innerType === t.Number ? Textbox.numberTransformer :
-      Textbox.transformer;
+    return options.transformer
+      ? options.transformer
+      : this.typeInfo.innerType === t.Number
+        ? Textbox.numberTransformer
+        : Textbox.transformer;
   }
 
   getTemplate() {
@@ -242,7 +263,8 @@ class Textbox extends Component {
     locals.placeholder = this.getPlaceholder();
     locals.onChangeNative = this.props.options.onChange;
     locals.keyboardType = this.getKeyboardType();
-    locals.underlineColorAndroid = (this.props.options.underlineColorAndroid || 'transparent');
+    locals.underlineColorAndroid =
+      this.props.options.underlineColorAndroid || 'transparent';
 
     [
       'help',
@@ -272,53 +294,48 @@ class Textbox extends Component {
       'onKeyPress',
       'returnKeyType',
       'selectionState'
-    ].forEach((name) => locals[name] = this.props.options[name]);
+    ].forEach(name => (locals[name] = this.props.options[name]));
 
     return locals;
   }
-
 }
 
 Textbox.transformer = {
-  format: value => Nil.is(value) ? '' : value,
+  format: value => (Nil.is(value) ? '' : value),
   parse: toNull
 };
 
 Textbox.numberTransformer = {
-  format: value => Nil.is(value) ? '' : String(value),
+  format: value => (Nil.is(value) ? '' : String(value)),
   parse: parseNumber
 };
 
 class Checkbox extends Component {
-
   getTemplate() {
     return this.props.options.template || this.props.ctx.templates.checkbox;
   }
 
   getLocals() {
     const locals = super.getLocals();
-    locals.label = this.props.ctx.auto !== 'none' ? locals.label || this.getDefaultLabel() : null;
+    locals.label =
+      this.props.ctx.auto !== 'none'
+        ? locals.label || this.getDefaultLabel()
+        : null;
 
-    [
-      'help',
-      'disabled',
-      'onTintColor',
-      'thumbTintColor',
-      'tintColor'
-    ].forEach((name) => locals[name] = this.props.options[name]);
+    ['help', 'disabled', 'onTintColor', 'thumbTintColor', 'tintColor'].forEach(
+      name => (locals[name] = this.props.options[name])
+    );
 
     return locals;
   }
-
 }
 
 Checkbox.transformer = {
-  format: value => Nil.is(value) ? false : value,
+  format: value => (Nil.is(value) ? false : value),
   parse: value => value
 };
 
 class Select extends Component {
-
   getTransformer() {
     const options = this.props.options;
     if (options.transformer) {
@@ -332,7 +349,7 @@ class Select extends Component {
   }
 
   getNullOption() {
-    return this.props.options.nullOption || {value: '', text: '-'};
+    return this.props.options.nullOption || { value: '', text: '-' };
   }
 
   getEnum() {
@@ -341,7 +358,9 @@ class Select extends Component {
 
   getOptions() {
     const options = this.props.options;
-    const items = options.options ? options.options.slice() : getOptionsOfEnum(this.getEnum());
+    const items = options.options
+      ? options.options.slice()
+      : getOptionsOfEnum(this.getEnum());
     if (options.order) {
       items.sort(getComparator(options.order));
     }
@@ -356,28 +375,23 @@ class Select extends Component {
     const locals = super.getLocals();
     locals.options = this.getOptions();
 
-    [
-      'help',
-      'enabled',
-      'mode',
-      'prompt',
-      'itemStyle'
-    ].forEach((name) => locals[name] = this.props.options[name]);
+    ['help', 'enabled', 'mode', 'prompt', 'itemStyle'].forEach(
+      name => (locals[name] = this.props.options[name])
+    );
 
     return locals;
   }
-
 }
 
-Select.transformer = (nullOption) => {
+Select.transformer = nullOption => {
   return {
-    format: value => Nil.is(value) && nullOption ? nullOption.value : String(value),
-    parse: value => nullOption && nullOption.value === value ? null : value
+    format: value =>
+      Nil.is(value) && nullOption ? nullOption.value : String(value),
+    parse: value => (nullOption && nullOption.value === value ? null : value)
   };
 };
 
 class DatePicker extends Component {
-
   getTemplate() {
     return this.props.options.template || this.props.ctx.templates.datepicker;
   }
@@ -392,28 +406,26 @@ class DatePicker extends Component {
       'minuteInterval',
       'mode',
       'timeZoneOffsetInMinutes',
-      'onPress',
-    ].forEach((name) => locals[name] = this.props.options[name]);
+      'onPress'
+    ].forEach(name => (locals[name] = this.props.options[name]));
 
     return locals;
   }
-
 }
 
 DatePicker.transformer = {
-  format: value => Nil.is(value) ? new Date() : value,
+  format: value => (Nil.is(value) ? new Date() : value),
   parse: value => value
 };
 
 class Struct extends Component {
-
   isValueNully() {
-    return Object.keys(this.refs).every((ref) => this.refs[ref].isValueNully());
+    return Object.keys(this.refs).every(ref => this.refs[ref].isValueNully());
   }
 
   removeErrors() {
-    this.setState({hasError: false});
-    Object.keys(this.refs).forEach((ref) => this.refs[ref].removeErrors());
+    this.setState({ hasError: false });
+    Object.keys(this.refs).forEach(ref => this.refs[ref].removeErrors());
   }
 
   getValue() {
@@ -432,7 +444,7 @@ class Struct extends Component {
 
     if (this.typeInfo.isMaybe && this.isValueNully()) {
       this.removeErrors();
-      return new t.ValidationResult({errors: [], value: null});
+      return new t.ValidationResult({ errors: [], value: null });
     }
 
     for (const ref in this.refs) {
@@ -453,14 +465,14 @@ class Struct extends Component {
       }
     }
 
-    this.setState({hasError: hasError});
-    return new t.ValidationResult({errors, value});
+    this.setState({ hasError: hasError });
+    return new t.ValidationResult({ errors, value });
   }
 
   onChange(fieldName, fieldValue, path) {
     const value = t.mixin({}, this.state.value);
     value[fieldName] = fieldValue;
-    this.setState({value}, () => {
+    this.setState({ value }, () => {
       this.props.onChange(value, path);
     });
   }
@@ -485,7 +497,7 @@ class Struct extends Component {
     const { ctx, options } = this.props;
     const props = this.getTypeProps();
     const auto = this.getAuto();
-    const i18n =  this.getI18n();
+    const i18n = this.getI18n();
     const config = this.getConfig();
     const value = this.state.value || {};
     const templates = this.getTemplates();
@@ -497,26 +509,34 @@ class Struct extends Component {
         const propValue = value[prop];
         const propType = getTypeFromUnion(type, propValue);
         const fieldsOptions = options.fields || noobj;
-        const propOptions = getComponentOptions(fieldsOptions[prop], noobj, propValue, type);
-        inputs[prop] = React.createElement(getFormComponent(propType, propOptions), {
-          key: prop,
-          ref: prop,
-          type: propType,
-          options: propOptions,
-          value: value[prop],
-          onChange: this.onChange.bind(this, prop),
-          ctx: {
-            context: ctx.context,
-            uidGenerator: ctx.uidGenerator,
-            auto,
-            config,
-            label: humanize(prop),
-            i18n,
-            stylesheet,
-            templates,
-            path: this.props.ctx.path.concat(prop)
+        const propOptions = getComponentOptions(
+          fieldsOptions[prop],
+          noobj,
+          propValue,
+          type
+        );
+        inputs[prop] = React.createElement(
+          getFormComponent(propType, propOptions),
+          {
+            key: prop,
+            ref: prop,
+            type: propType,
+            options: propOptions,
+            value: value[prop],
+            onChange: this.onChange.bind(this, prop),
+            ctx: {
+              context: ctx.context,
+              uidGenerator: ctx.uidGenerator,
+              auto,
+              config,
+              label: humanize(prop),
+              i18n,
+              stylesheet,
+              templates,
+              path: this.props.ctx.path.concat(prop)
+            }
           }
-        });
+        );
       }
     }
     return inputs;
@@ -530,7 +550,6 @@ class Struct extends Component {
     locals.template = templates.struct;
     return locals;
   }
-
 }
 
 function toSameLength(value, keys, uidGenerator) {
@@ -538,14 +557,13 @@ function toSameLength(value, keys, uidGenerator) {
     return keys;
   }
   const ret = [];
-  for (let i = 0, len = value.length; i < len; i++ ) {
+  for (let i = 0, len = value.length; i < len; i++) {
     ret[i] = keys[i] || uidGenerator.next();
   }
   return ret;
 }
 
 export class List extends Component {
-
   constructor(props) {
     super(props);
     this.state.keys = this.state.value.map(() => props.ctx.uidGenerator.next());
@@ -568,12 +586,12 @@ export class List extends Component {
 
   removeErrors() {
     this.setState({ hasError: false });
-    Object.keys(this.refs).forEach((ref) => this.refs[ref].removeErrors());
+    Object.keys(this.refs).forEach(ref => this.refs[ref].removeErrors());
   }
 
   getValue() {
     const value = [];
-    for (let i = 0, len = this.state.value.length; i < len; i++ ) {
+    for (let i = 0, len = this.state.value.length; i < len; i++) {
       if (this.refs.hasOwnProperty(i)) {
         value.push(this.refs[i].getValue());
       }
@@ -589,10 +607,10 @@ export class List extends Component {
 
     if (this.typeInfo.isMaybe && this.isValueNully()) {
       this.removeErrors();
-      return new t.ValidationResult({errors: [], value: null});
+      return new t.ValidationResult({ errors: [], value: null });
     }
 
-    for (let i = 0, len = this.state.value.length; i < len; i++ ) {
+    for (let i = 0, len = this.state.value.length; i < len; i++) {
       result = this.refs[i].validate();
       errors = errors.concat(result.errors);
       value.push(result.value);
@@ -605,8 +623,8 @@ export class List extends Component {
       errors = errors.concat(result.errors);
     }
 
-    this.setState({hasError: hasError});
-    return new t.ValidationResult({errors: errors, value: value});
+    this.setState({ hasError: hasError });
+    return new t.ValidationResult({ errors: errors, value: value });
   }
 
   onChange(value, keys, path, kind) {
@@ -619,7 +637,12 @@ export class List extends Component {
   addItem() {
     const value = this.state.value.concat(undefined);
     const keys = this.state.keys.concat(this.props.ctx.uidGenerator.next());
-    this.onChange(value, keys, this.props.ctx.path.concat(value.length - 1), 'add');
+    this.onChange(
+      value,
+      keys,
+      this.props.ctx.path.concat(value.length - 1),
+      'add'
+    );
   }
 
   onItemChange(itemIndex, itemValue, path, kind) {
@@ -677,7 +700,12 @@ export class List extends Component {
     return value.map((itemValue, i) => {
       const type = this.typeInfo.innerType.meta.type;
       const itemType = getTypeFromUnion(type, itemValue);
-      const itemOptions = getComponentOptions(options.item, noobj, itemValue, type);
+      const itemOptions = getComponentOptions(
+        options.item,
+        noobj,
+        itemValue,
+        type
+      );
       const ItemComponent = getFormComponent(itemType, itemOptions);
       const buttons = [];
       if (!options.disableRemove) {
@@ -688,16 +716,18 @@ export class List extends Component {
         });
       }
       if (!options.disableOrder) {
-        buttons.push({
-          type: 'move-up',
-          label: i18n.up,
-          click: this.moveUpItem.bind(this, i)
-        },
-        {
-          type: 'move-down',
-          label: i18n.down,
-          click: this.moveDownItem.bind(this, i)
-        });
+        buttons.push(
+          {
+            type: 'move-up',
+            label: i18n.up,
+            click: this.moveUpItem.bind(this, i)
+          },
+          {
+            type: 'move-down',
+            label: i18n.down,
+            click: this.moveDownItem.bind(this, i)
+          }
+        );
       }
       return {
         input: React.createElement(ItemComponent, {
@@ -728,25 +758,25 @@ export class List extends Component {
     const options = this.props.options;
     const i18n = this.getI18n();
     const locals = super.getLocals();
-    locals.add = options.disableAdd ? null : {
-      type: 'add',
-      label: i18n.add,
-      click: this.addItem.bind(this)
-    };
+    locals.add = options.disableAdd
+      ? null
+      : {
+          type: 'add',
+          label: i18n.add,
+          click: this.addItem.bind(this)
+        };
     locals.items = this.getItems();
     locals.className = options.className;
     return locals;
   }
-
 }
 
 List.transformer = {
-  format: value => Nil.is(value) ? noarr : value,
+  format: value => (Nil.is(value) ? noarr : value),
   parse: value => value
 };
 
 class Form extends React.Component {
-
   pureValidate() {
     return this.refs.input.pureValidate();
   }
@@ -787,14 +817,21 @@ class Form extends React.Component {
   }
 
   render() {
-
     const stylesheet = this.props.stylesheet || Form.stylesheet;
     const templates = this.props.templates || Form.templates;
     const i18n = this.props.i18n || Form.i18n;
 
     if (process.env.NODE_ENV !== 'production') {
-      t.assert(t.isType(this.props.type), `[${SOURCE}] missing required prop type`);
-      t.assert(t.maybe(t.Object).is(this.props.options) || t.Function.is(this.props.options) || t.list(t.maybe(t.Object)).is(this.props.options), `[${SOURCE}] prop options, if specified, must be an object, a function returning the options or a list of options for unions`);
+      t.assert(
+        t.isType(this.props.type),
+        `[${SOURCE}] missing required prop type`
+      );
+      t.assert(
+        t.maybe(t.Object).is(this.props.options) ||
+          t.Function.is(this.props.options) ||
+          t.list(t.maybe(t.Object)).is(this.props.options),
+        `[${SOURCE}] prop options, if specified, must be an object, a function returning the options or a list of options for unions`
+      );
       t.assert(t.Object.is(stylesheet), `[${SOURCE}] missing stylesheet config`);
       t.assert(t.Object.is(templates), `[${SOURCE}] missing templates config`);
       t.assert(t.Object.is(i18n), `[${SOURCE}] missing i18n config`);
@@ -802,7 +839,12 @@ class Form extends React.Component {
 
     const value = this.props.value;
     const type = getTypeFromUnion(this.props.type, value);
-    const options = getComponentOptions(this.props.options, noobj, value, this.props.type);
+    const options = getComponentOptions(
+      this.props.options,
+      noobj,
+      value,
+      this.props.type
+    );
 
     // this is in the render method because I need this._reactInternalInstance._rootNodeID in React ^0.14.0
     // and this._reactInternalInstance._nativeContainerInfo._idCounter in React ^15.0.0
@@ -825,7 +867,6 @@ class Form extends React.Component {
       }
     });
   }
-
 }
 
 module.exports = {
